Added util functions from dashboard
